feat(material/legacy-progress-bar): step 4. move mdc folder
