+++ src/dev-app/mdc-progress-bar/BUILD.bazel
@@ -11,8 +11,8 @@ ng_module(
     ],
     deps = [
         "//src/material-experimental/mdc-button",
-        "//src/material-experimental/mdc-progress-bar",
         "//src/material/button-toggle",
+        "//src/material/progress-bar",
         "@npm//@angular/forms",
     ],
 )
