+++ src/e2e-app/BUILD.bazel
@@ -48,7 +48,6 @@ ng_module(
         "//src/material-experimental/mdc-dialog",
         "//src/material-experimental/mdc-input",
         "//src/material-experimental/mdc-menu",
-        "//src/material-experimental/mdc-progress-bar",
         "//src/material-experimental/mdc-progress-spinner",
         "//src/material-experimental/mdc-radio",
         "//src/material-experimental/mdc-slide-toggle",
@@ -67,6 +66,7 @@ ng_module(
         "//src/material/legacy-progress-bar",
         "//src/material/list",
         "//src/material/menu",
+        "//src/material/progress-bar",
         "//src/material/progress-spinner",
         "//src/material/radio",
         "//src/material/select",
