+++ src/material-experimental/mdc-core/typography/_all-typography.import.scss
@@ -75,7 +75,6 @@ $mat-mdc-mdc-text-field-outlined-idle-border, $mat-mdc-mdc-text-field-placeholde
 @forward '../../mdc-list/list-theme' as mat-mdc-list-*;
 @forward '../../mdc-menu/menu-theme' as mat-mdc-menu-*;
 @forward '../../mdc-paginator/paginator-theme' as mat-mdc-paginator-*;
-@forward '../../mdc-progress-bar/progress-bar-theme' as mat-mdc-progress-bar-*;
 @forward '../../mdc-progress-spinner/progress-spinner-theme' as mat-mdc-progress-spinner-*;
 @forward '../../mdc-input/input-theme' as mat-mdc-input-*;
 @forward '../../mdc-form-field/form-field-density' as mat-mdc-*;
