+++ src/material/config.bzl
@@ -42,6 +42,8 @@ entryPoints = [
     "paginator/testing",
     "legacy-progress-bar",
     "legacy-progress-bar/testing",
+    "progress-bar",
+    "progress-bar/testing",
     "progress-spinner",
     "progress-spinner/testing",
     "radio",
