+++ src/material/progress-bar/BUILD.bazel
@@ -1,6 +1,7 @@
 load("//src/e2e-app:test_suite.bzl", "e2e_test_suite")
 load(
     "//tools:defaults.bzl",
+    "markdown_to_html",
     "ng_e2e_test_library",
     "ng_module",
     "ng_test_library",
@@ -12,7 +13,7 @@ load(
 package(default_visibility = ["//visibility:public"])
 
 ng_module(
-    name = "mdc-progress-bar",
+    name = "progress-bar",
     srcs = glob(
         ["**/*.ts"],
         exclude = [
@@ -21,18 +22,17 @@ ng_module(
     ),
     assets = [":progress_bar_scss"] + glob(["**/*.html"]),
     deps = [
-        "//src/material-experimental/mdc-core",
+        "//src/material/core",
         "//src/material/legacy-progress-bar",
         "@npm//@angular/core",
     ],
 )
 
 sass_library(
-    name = "mdc_progress_bar_scss_lib",
+    name = "progress_bar_scss_lib",
     srcs = glob(["**/_*.scss"]),
     deps = [
         "//:mdc_sass_lib",
-        "//src/material:sass_lib",
         "//src/material/core:core_scss_lib",
     ],
 )
@@ -42,7 +42,6 @@ sass_binary(
     src = "progress-bar.scss",
     deps = [
         "//:mdc_sass_lib",
-        "//src/material:sass_lib",
         "//src/material/core:core_scss_lib",
     ],
 )
@@ -54,7 +53,7 @@ ng_test_library(
         exclude = ["**/*.e2e.spec.ts"],
     ),
     deps = [
-        ":mdc-progress-bar",
+        ":progress-bar",
         "//src/cdk/testing/private",
         "@npm//@angular/platform-browser",
     ],
@@ -81,4 +80,14 @@ e2e_test_suite(
         ":e2e_test_sources",
         "//src/cdk/testing/private/e2e",
     ],
+)
+
+markdown_to_html(
+    name = "overview",
+    srcs = [":progress-bar.md"],
+)
+
+filegroup(
+    name = "source-files",
+    srcs = glob(["**/*.ts"]),
 )
