+++ src/material/progress-bar/_progress-bar-theme.scss
@@ -1,4 +1,5 @@
-@use '@angular/material' as mat;
+@use '../core/mdc-helpers/mdc-helpers';
+@use '../core/theming/theming';
 @use '@material/theme/theme-color' as mdc-theme-color;
 @use '@material/linear-progress' as mdc-linear-progress;
 @use 'sass:color';
@@ -10,14 +11,14 @@
   // and doesn't match the Material design spec. For now we approximate the buffer background by
   // applying an opacity to the color of the bar.
   $buffer-color: color.adjust(mdc-theme-color.prop-value($color), $alpha: -0.75);
-  @include mdc-linear-progress.bar-color($color, $query: mat.$private-mdc-theme-styles-query);
+  @include mdc-linear-progress.bar-color($color, $query: mdc-helpers.$mdc-theme-styles-query);
   @include mdc-linear-progress.buffer-color($buffer-color,
-    $query: mat.$private-mdc-theme-styles-query);
+    $query: mdc-helpers.$mdc-theme-styles-query);
 }
 
 @mixin color($config-or-theme) {
-  $config: mat.get-color-config($config-or-theme);
-  @include mat.private-using-mdc-theme($config) {
+  $config: theming.get-color-config($config-or-theme);
+  @include mdc-helpers.using-mdc-theme($config) {
     .mat-mdc-progress-bar {
       @include _palette-styles(primary);
 
@@ -37,11 +38,11 @@
 @mixin density($config-or-theme) {}
 
 @mixin theme($theme-or-color-config) {
-  $theme: mat.private-legacy-get-theme($theme-or-color-config);
-  @include mat.private-check-duplicate-theme-styles($theme, 'mat-mdc-progress-bar') {
-    $color: mat.get-color-config($theme);
-    $density: mat.get-density-config($theme);
-    $typography: mat.get-typography-config($theme);
+  $theme: theming.private-legacy-get-theme($theme-or-color-config);
+  @include theming.private-check-duplicate-theme-styles($theme, 'mat-progress-bar') {
+    $color: theming.get-color-config($theme);
+    $density: theming.get-density-config($theme);
+    $typography: theming.get-typography-config($theme);
 
     @if $color != null {
       @include color($color);
