+++ src/material/progress-bar/progress-bar.scss
@@ -1,8 +1,8 @@
-@use '@angular/material' as mat;
+@use '../core/mdc-helpers/mdc-helpers';
 @use '@material/linear-progress' as mdc-linear-progress;
 
-@include mat.private-disable-mdc-fallback-declarations {
-  @include mdc-linear-progress.core-styles($query: mat.$private-mdc-base-styles-query);
+@include mdc-helpers.disable-mdc-fallback-declarations {
+  @include mdc-linear-progress.core-styles($query: mdc-helpers.$mdc-base-styles-query);
 }
 
 .mat-mdc-progress-bar {
