+++ src/material/progress-bar/progress-bar.ts
@@ -21,7 +21,7 @@ import {
   OnDestroy,
   ChangeDetectorRef,
 } from '@angular/core';
-import {CanColor, mixinColor} from '@angular/material-experimental/mdc-core';
+import {CanColor, mixinColor} from '@angular/material/core';
 import {ANIMATION_MODULE_TYPE} from '@angular/platform-browser/animations';
 import {
   MatLegacyProgressBarDefaultOptions,
