+++ src/material/progress-bar/public-api.ts
