+++ src/material/progress-bar/testing/progress-bar-harness.spec.ts
