+++ src/material/progress-bar/testing/public-api.ts
