+++ tools/public_api_guard/material/progress-bar-testing.md
@@ -1,0 +1,25 @@
+## API Report File for "components-srcs"
+
+> Do not edit this file. It is a report generated by [API Extractor](https://api-extractor.com/).
+
+```ts
+
+import { ComponentHarness } from '@angular/cdk/testing';
+import { ComponentHarnessConstructor } from '@angular/cdk/testing';
+import { HarnessPredicate } from '@angular/cdk/testing';
+import { LegacyProgressBarHarnessFilters as ProgressBarHarnessFilters } from '@angular/material/legacy-progress-bar/testing';
+
+// @public
+export class MatProgressBarHarness extends ComponentHarness {
+    getMode(): Promise<string | null>;
+    getValue(): Promise<number | null>;
+    // (undocumented)
+    static hostSelector: string;
+    static with<T extends MatProgressBarHarness>(this: ComponentHarnessConstructor<T>, options?: ProgressBarHarnessFilters): HarnessPredicate<T>;
+}
+
+export { ProgressBarHarnessFilters }
+
+// (No @packageDocumentation comment for this package)
+
+```
