+++ tools/public_api_guard/material/progress-bar.md
@@ -1,0 +1,78 @@
+## API Report File for "components-srcs"
+
+> Do not edit this file. It is a report generated by [API Extractor](https://api-extractor.com/).
+
+```ts
+
+import { _AbstractConstructor } from '@angular/material/core';
+import { AfterViewInit } from '@angular/core';
+import { CanColor } from '@angular/material/core';
+import { ChangeDetectorRef } from '@angular/core';
+import { _Constructor } from '@angular/material/core';
+import { ElementRef } from '@angular/core';
+import { EventEmitter } from '@angular/core';
+import * as i0 from '@angular/core';
+import * as i2 from '@angular/material/core';
+import { MAT_LEGACY_PROGRESS_BAR_DEFAULT_OPTIONS as MAT_PROGRESS_BAR_DEFAULT_OPTIONS } from '@angular/material/legacy-progress-bar';
+import { MAT_LEGACY_PROGRESS_BAR_LOCATION as MAT_PROGRESS_BAR_LOCATION } from '@angular/material/legacy-progress-bar';
+import { MAT_LEGACY_PROGRESS_BAR_LOCATION_FACTORY as MAT_PROGRESS_BAR_LOCATION_FACTORY } from '@angular/material/legacy-progress-bar';
+import { MatLegacyProgressBarDefaultOptions as MatProgressBarDefaultOptions } from '@angular/material/legacy-progress-bar';
+import { MatLegacyProgressBarLocation as MatProgressBarLocation } from '@angular/material/legacy-progress-bar';
+import { NgZone } from '@angular/core';
+import { OnDestroy } from '@angular/core';
+import { LegacyProgressAnimationEnd as ProgressAnimationEnd } from '@angular/material/legacy-progress-bar';
+
+export { MAT_PROGRESS_BAR_DEFAULT_OPTIONS }
+
+export { MAT_PROGRESS_BAR_LOCATION }
+
+export { MAT_PROGRESS_BAR_LOCATION_FACTORY }
+
+// @public (undocumented)
+export class MatProgressBar extends _MatProgressBarBase implements AfterViewInit, OnDestroy, CanColor {
+    constructor(elementRef: ElementRef<HTMLElement>, _ngZone: NgZone, _changeDetectorRef: ChangeDetectorRef, _animationMode?: string | undefined, defaults?: MatProgressBarDefaultOptions);
+    readonly animationEnd: EventEmitter<ProgressAnimationEnd>;
+    // (undocumented)
+    _animationMode?: string | undefined;
+    get bufferValue(): number;
+    set bufferValue(v: number);
+    _getBufferBarFlexBasis(): string;
+    _getPrimaryBarTransform(): string;
+    _isIndeterminate(): boolean;
+    _isNoopAnimation: boolean;
+    get mode(): ProgressBarMode;
+    set mode(value: ProgressBarMode);
+    // (undocumented)
+    ngAfterViewInit(): void;
+    // (undocumented)
+    ngOnDestroy(): void;
+    get value(): number;
+    set value(v: number);
+    // (undocumented)
+    static ɵcmp: i0.ɵɵComponentDeclaration<MatProgressBar, "mat-progress-bar", ["matProgressBar"], { "color": "color"; "value": "value"; "bufferValue": "bufferValue"; "mode": "mode"; }, { "animationEnd": "animationEnd"; }, never, never, false>;
+    // (undocumented)
+    static ɵfac: i0.ɵɵFactoryDeclaration<MatProgressBar, [null, null, null, { optional: true; }, { optional: true; }]>;
+}
+
+export { MatProgressBarDefaultOptions }
+
+export { MatProgressBarLocation }
+
+// @public (undocumented)
+export class MatProgressBarModule {
+    // (undocumented)
+    static ɵfac: i0.ɵɵFactoryDeclaration<MatProgressBarModule, never>;
+    // (undocumented)
+    static ɵinj: i0.ɵɵInjectorDeclaration<MatProgressBarModule>;
+    // (undocumented)
+    static ɵmod: i0.ɵɵNgModuleDeclaration<MatProgressBarModule, [typeof i1.MatProgressBar], never, [typeof i1.MatProgressBar, typeof i2.MatCommonModule]>;
+}
+
+export { ProgressAnimationEnd }
+
+// @public (undocumented)
+export type ProgressBarMode = 'determinate' | 'indeterminate' | 'buffer' | 'query';
+
+// (No @packageDocumentation comment for this package)
+
+```
